refactor: organize enums and ID value objects into dedicated folders
- Moved all enums to a new folder named "Enums" for improved structure
- Created an "Ids" folder within "ValueObjects" and moved ID value objects there
