+++ src/SMAIAXBackend.Domain/Model/Entities/Measurement.cs
@@ -1,4 +1,5 @@
 using SMAIAXBackend.Domain.Model.ValueObjects;
+using SMAIAXBackend.Domain.Model.ValueObjects.Ids;
 
 namespace SMAIAXBackend.Domain.Model.Entities;
