+++ src/SMAIAXBackend.Domain/Model/Entities/Policy.cs
@@ -1,4 +1,6 @@
+using SMAIAXBackend.Domain.Model.Enums;
 using SMAIAXBackend.Domain.Model.ValueObjects;
+using SMAIAXBackend.Domain.Model.ValueObjects.Ids;
 
 namespace SMAIAXBackend.Domain.Model.Entities;
