+++ src/SMAIAXBackend.Domain/Model/Entities/SmartMeter.cs
@@ -1,4 +1,4 @@
-using SMAIAXBackend.Domain.Model.ValueObjects;
+using SMAIAXBackend.Domain.Model.ValueObjects.Ids;
 
 namespace SMAIAXBackend.Domain.Model.Entities;
