+++ src/SMAIAXBackend.Domain/Model/ValueObjects/Ids/RefreshTokenId.cs
@@ -1,6 +1,6 @@
-namespace SMAIAXBackend.Domain.Model.ValueObjects;
+namespace SMAIAXBackend.Domain.Model.ValueObjects.Ids;
 
-public class MetadataId(Guid id) : ValueObject
+public class RefreshTokenId(Guid id) : ValueObject
 {
     public Guid Id { get; } = id;
