+++ src/SMAIAXBackend.Infrastructure/EntityConfigurations/DomainUserConfiguration.cs
@@ -2,7 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 using SMAIAXBackend.Domain.Model.Entities;
-using SMAIAXBackend.Domain.Model.ValueObjects;
+using SMAIAXBackend.Domain.Model.ValueObjects.Ids;
 
 namespace SMAIAXBackend.Infrastructure.EntityConfigurations;
