+++ tests/SMAIAXBackend.Domain.UnitTests/PolicyRequestTests.cs
@@ -1,7 +1,9 @@
 using System.Globalization;
 
 using SMAIAXBackend.Domain.Model.Entities;
+using SMAIAXBackend.Domain.Model.Enums;
 using SMAIAXBackend.Domain.Model.ValueObjects;
+using SMAIAXBackend.Domain.Model.ValueObjects.Ids;
 
 namespace SMAIAXBackend.Domain.UnitTests;
